requirements.txt
Django==2.0.2
djangorestframework==3.7.7
django-cors-headers==2.1.0
django-filter==1.1.0
coreapi==2.3.3
factory-boy==2.10.0
Markdown==2.6.11
Pygments==2.2.0
imailing==1.0.0
Pillow==5.1.0
python-decouple==3.1
django-storages==1.6.6
